requirements.txt
Django == 4.2.7
psycopg == 3.1.13
pyjwt == 2.8.0
django-cors-headers == 4.3.1
djangorestframework==3.14.0
